Create  outside the component in a separate file
