+++ src/logic/event.js
@@ -1,0 +1,95 @@
+import React, { useState, useEffect } from "react";
+import NumberFormat from "react-number-format";
+
+const event = () => {
+  const [preState, setPreState] = useState("");
+  const [curState, setCurState] = useState("");
+  const [input, setInput] = useState("0");
+  const [operator, setOperator] = useState(null);
+  const [total, setTotal] = useState(false);
+
+  const inputNum = (e) => {
+    if (curState.includes(".") && e.target.innerText === ".") return;
+
+    if (total) {
+      setPreState("");
+    }
+    if (curState) {
+      setCurState((pre) => pre + e.target.innerText);
+    } else {
+      setCurState(e.target.innerText);
+      setTotal(false);
+    }
+  };
+
+  useEffect(() => {
+    setInput(curState);
+  }, [curState]);
+
+  useEffect(() => {
+    setInput("0");
+  }, []);
+
+  // equals
+
+  const equals = (e) => {
+    if (e?.target.innerText === "=") {
+      setTotal(true);
+    }
+    let cal;
+    switch (operator) {
+      case "÷":
+        cal = String(parseFloat(preState) / parseFloat(curState));
+        break;
+
+      case "+":
+        cal = String(parseFloat(preState) + parseFloat(curState));
+        break;
+      case "X":
+        cal = String(parseFloat(preState) * parseFloat(curState));
+        break;
+      case "-":
+        cal = String(parseFloat(preState) - parseFloat(curState));
+        break;
+      default:
+        return;
+    }
+    setInput("");
+    setPreState(cal);
+    setCurState("");
+  };
+
+  const operatorType = (e) => {
+    setTotal(false);
+    setOperator(e.target.innerText);
+    if (curState === "") return;
+    if (preState !== "") {
+      equals();
+    } else {
+      setPreState(curState);
+      setCurState("");
+    }
+  };
+
+  const minusPlus = () => {
+    if (curState.charAt(0) === "-") {
+      setCurState(curState.substring(1));
+    } else {
+      setCurState(`-${curState}`);
+    }
+  };
+
+  const percent = () => {
+    if (preState) {
+      setCurState(String((parseFloat(curState) / 100) * preState));
+    } else {
+      setCurState(String(parseFloat(curState) / 100));
+    }
+  };
+
+  const reset = () => {
+    setInput("0");
+    setPreState("");
+    setCurState("");
+  };
+};
